[Update] Refactor stats - moved to a separate app
