+++ mask_detect/mask_detect/urls.py
@@ -19,6 +19,7 @@ from django.urls import path
 from django.conf import settings
 from django.conf.urls.static import static
 from accounts import views as accounts_views
+from stats import views as stat_views
 
 
 urlpatterns = [
@@ -28,9 +29,9 @@ urlpatterns = [
     path('login/', auth_views.LoginView.as_view(template_name='accounts/login.html'), name="login"),
     path('logout/', auth_views.LogoutView.as_view(template_name='accounts/logout.html'), name="logout"),
     path('profile/', accounts_views.profile, name="profile"),
-    path('stats', accounts_views.export_csv_stats, name='export-csv'),
-    path('dashboard/', accounts_views.dashboard, name="dashboard"),
-    path('admin-stats/', accounts_views.dashboard_export_csvs, name='export-all-stats')
+    path('stats', stat_views.export_csv_stats, name='export-csv'),
+    path('dashboard/', stat_views.dashboard, name="dashboard"),
+    path('admin-stats/', stat_views.dashboard_export_csvs, name='export-all-stats')
 ]
