+++ mask_detect/stats/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Statistic
+
+admin.site.register(Statistic)
